Utils/requirements.txt
colorama==0.4.6
coverage==7.6.1
parameterized==0.9.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-soft-assertions==0.1.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
webdriver-manager==4.0.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest-html==4.1.1

